backend/requirements.txt
Django == 4.1.2
djangorestframework == 3.14.0
django-cors-headers == 3.13.0
drf-writable-nested == 0.7.0
drf-extra-fields == 3.4.1
django-environ == 0.9.0
django-nested-inline~=0.4.5
sqlparse==0.4.3
Pillow==9.3.0
